requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-environ==0.12.0
gunicorn==23.0.0
whitenoise==6.11.0
djangorestframework-simplejwt==5.5.1
